requirements.txt
python-telegram-bot==20.7
openai==1.3.7
python-dotenv==1.0.0
sqlalchemy>=2.0.25
alembic==1.12.1
# psycopg2-binary==2.9.9  # Requires PostgreSQL headers, uncomment if using PostgreSQL
redis==5.0.1
celery==5.3.4
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
linkedin-api==2.0.0
facebook-sdk==3.1.0
instagram-private-api==1.6.0
pydantic>=2.6.0
pydantic-settings>=2.0.0
fastapi==0.104.1
uvicorn==0.24.0
pytest==7.4.3
pytest-asyncio==0.21.1
setuptools==58.2.0